Mail/requirements.txt
gspread>=5.12.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
